# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HuntFlow vNext API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/match-scores/run:
    post:
      summary: Score candidates for a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchScoreRequest"
      responses:
        "200":
          description: Scoring result
  /api/v1/submissions/draft:
    post:
      summary: Generate a recommendation draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionDraftRequest"
      responses:
        "201":
          description: Draft created
  /api/v1/submissions:
    get:
      summary: List recommendation drafts and submissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Submission list
  /api/v1/submissions/{submission_id}:
    get:
      summary: Fetch a submission draft
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submission_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submission detail
  /api/v1/submissions/{submission_id}/submit:
    post:
      summary: Submit an approved draft using an approval token
      description: |
        Final formal submission step.
        Flow is strict: request approval -> approve -> submit with `approval_token`.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submission_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Submit the approved draft with the approval token returned by the approval flow.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionSubmitRequest"
      responses:
        "200":
          description: Submission submitted
        "400":
          description: Approval token is required
        "403":
          description: Approval token is invalid
        "404":
          description: Submission not found
        "409":
          description: Submission is not in DRAFT state
...
  /api/v1/source-adapters:
    get:
      summary: List public controlled source adapters (structured-import only)
      description: Browser prototype capture aliases are intentionally excluded from the public experimental contract.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Source adapter list
  /api/v1/source-runs:
    post:
      summary: Create a buffered controlled sourcing run from structured-import
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceRunRequest"
      responses:
        "201":
          description: Source run created
    get:
      summary: List source runs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Source run list
  /api/v1/source-runs/{run_id}:
    get:
      summary: Get a source run and its raw items
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Source run detail
  /api/v1/source-items/{item_id}/review:
    post:
      summary: Review a buffered source item with a single terminal decision
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: item_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceReviewRequest"
      responses:
        "200":
          description: Review recorded
  /api/v1/source-items/{item_id}/promote:
    post:
      summary: Promote an approved buffered source item into candidates and pipelines
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: item_id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Item promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcePromoteResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: huntflow_session
  schemas:
...
    MatchScoreRequest:
      type: object
      required: [job_order_id, candidate_ids]
      properties:
        job_order_id:
          type: string
        candidate_ids:
          type: array
          items:
            type: string
    SubmissionDraftRequest:
      type: object
      required: [job_order_id, candidate_id]
      properties:
        job_order_id:
          type: string
        candidate_id:
          type: string
        include_gap_analysis:
          type: boolean
    SubmissionSubmitRequest:
      type: object
      required: [approval_token]
      description: Approval-gated submit payload.
      properties:
        approval_token:
          type: string
          minLength: 1
          description: Token from an APPROVED submission approval decision.
...
    SourceRunRequest:
      type: object
      required: [job_order_id, source_name]
      description: Public experimental lane only supports structured-import review/promotion flow.
      properties:
        job_order_id:
          type: string
        source_name:
          type: string
          enum: [structured-import]
          description: Public experimental contract only accepts structured-import.
        items:
          type: array
          items:
            $ref: "#/components/schemas/SourceItemPayload"
        source_config:
          $ref: "#/components/schemas/SourceConfigPayload"
    SourceItemPayload:
      type: object
      properties:
        full_name:
          type: string
        current_company:
          type: string
        current_title:
          type: string
        city:
          type: string
        email:
          type: string
        phone:
          type: string
        resume_text:
          type: string
        summary:
          type: string
        source_url:
          type: string
        capture_notes:
          type: string
    SourceConfigPayload:
      type: object
      properties:
        source_url:
          type: string
        source_label:
          type: string
        capture_notes:
          type: string
        candidate_name:
          type: string
        current_company:
          type: string
        current_title:
          type: string
        city:
          type: string
        email:
          type: string
        phone:
          type: string
        resume_text:
          type: string
        captured_at:
          type: string
    SourceReviewRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        note:
          type: string
...
    CandidateSummary:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        current_company:
          type: string
        current_title:
          type: string
        city:
          type: string
        source_type:
          type: string
        email_masked:
          type: string
          nullable: true
        phone_masked:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    SourcePromoteResponse:
      type: object
      properties:
        candidate:
          $ref: "#/components/schemas/CandidateSummary"
        pipeline_id:
          type: string
          nullable: true
        source_item_id:
          type: string